requirements.txt
# Base de datos y ORM
SQLAlchemy==2.0.23
mysql-connector-python==8.2.0
pymysql==1.1.0
alembic==1.13.0
redis==5.0.1

# Web Scraping y Automatización
praw==7.7.1                    # Para la API de Reddit
playwright==1.40.0             # Añadido para TikTok upload automation
selenium==4.15.2               # Para scraping
webdriver-manager==4.0.1
fake-useragent==1.4.0
requests==2.31.0
urllib3==2.0.7

# Procesamiento de Audio y Texto
edge-tts==6.1.7               # Añadido para TTS
python-dateutil==2.8.2
langdetect==1.0.9
srt==3.5.3                    # Añadido para subtítulos

# Procesamiento de Video e Imágenes
moviepy==1.0.3
Pillow==10.1.0
opencv-python==4.8.1.78
ffmpeg-python==0.2.0
numpy==1.24.3                 # Añadido para procesamiento de video
scipy==1.10.1                 # Añadido para efectos de video

# Logging y Monitoreo
structlog==23.1.0
python-json-logger==2.0.7
colorama==0.4.6

# Utilidades y Herramientas
python-dotenv==1.0.0
tenacity==8.2.3               # Para reintentos
pathlib==1.0.1               # Añadido para manejo de paths
tqdm==4.66.1                 # Para barras de progreso
schedule==1.2.1              # Para programación de tareas

# Testing y Calidad de Código
pytest==8.0.0
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Asincronía
aiohttp==3.9.1
asyncio==3.4.3

# Desarrollo y Debug
ipython==8.12.0
ipdb==0.13.13
